Update Bolts to 1.5.0
We have to use version 1.4.1-SNAPSHOT of the transitive dependencies
`bolts-tasks` and `bolts-applinks`. These were not released as 1.5.0
but only as 1.4.1-SNAPSHOT and are functionally identical to the code
in version 1.5.0.

I compared the differences of `bolts-tasks` (we don't use
`bolts-applinks`) between tags 1.4.0 and 1.5.0 of the
`bolts-android` repo (https://github.com/BoltsFramework/Bolts-Android)
with the differences of `bolts-tasks` between tag 1.4.0 and the source
of 1.4.1-SNAPSHOT
(https://oss.sonatype.org/#nexus-search;quick~bolts-tasks).

The only difference is that in 1.5.0 the license headers were update,
otherwise 1.4.1-SNAPSHOT is the same as 1.5.0.